Split table format from query engine.
This commit aims to table format database engines (e.g. Iceberg) from query engines (e.g. Snowflake). The goal is to separate those concerns and support additional engines in a modular architecture (e.g. DeltaLake & Hudi for table formats and Spark, Flink, Athena, etc for query engines).

The concerns we are separating are:

- Table Formats: Provide the sink for Flink and the registration of the table DDLs in the configured catalog. It optimizes the DDL (i.e. partition, sort-order, and update strategy). The optimization is not implemented in this commit but future work.
- Query Engines: Define how to import a table from the catalog via DDL and the translated queries to execute over those tables.

The physical plan returned by a table format database engine is thus composite: it contains the DDL and catalog registration and the physical plans for all registered query engines.

The PipelineFactory registers all query engines with the table format databases that support them.
The engine type "QUERY" was added but it is only used for that purpose and is not part of the pipeline.

These changes have not been tested and aren't completed yet:
1) I was unsure of where and how I should separate the Iceberg side of the catalog registration from the snowflake side of the import. I think that's all in the IcebergDDLFactory and needs to be separated out.
2) Some aspects look hardcoded (e.g. the catalog) and hence I wasn't sure how this should be separated.

Signed-off-by: Matthias Broecheler <mbroecheler@users.noreply.github.com>