adding test coverage for timestamp arguments with interval conditions (#1239)
adding test coverage for timestamp arguments with interval conditions. For now, those intervals have to be on the side of the column, not the parameter. There is a limitation in Apache Calcite's parameter inference. We can look into adding explicit casts for the RexParam but that's a bigger change. For now, make sure the parameter is not part of a flexibly typed expression.

Signed-off-by: Matthias Broecheler <mbroecheler@users.noreply.github.com>