Streamline CLI params (#1695)
This required to make our internal test framework support running projects from a `package.json` in all of our test scenarios.

As the `--api` CLI option got removed as well, now if the `package.json` does not have a GraphQL schema definition, the inferred schema is written out to the build folder.